test: split test_local_cache.py by topic into its own scoped test package
